Major propagation rewrite, other bugfixes
